+ FileImageCache header, move implementation out of FileImage
