+++ mp/src/game/client/momentum/ui/controls/FileImage.cpp
@@ -9,6 +9,8 @@
 #include "mom_api_requests.h"
 #include "util/mom_util.h"
 
+#include "FileImageCache.h"
+
 #define STB_IMAGE_IMPLEMENTATION
 #define STBI_NO_STDIO
 #define STBI_NO_HDR
@@ -21,55 +23,8 @@
 
 using namespace vgui;
 
-struct ImageCacheEntry
-{
-    CUtlBuffer m_bufOriginalImage;
-};
-
-class CFileImageCache
-{
-public:
-    CFileImageCache() {}
-
-    void AddImageToCache(const char *pPath, const CUtlBuffer &pBuf)
-    {
-        const auto foundIndex = m_dictImages.Find(pPath);
-        if (m_dictImages.IsValidIndex(foundIndex))
-            return;
-
-        const auto pEntry = new ImageCacheEntry;
-        pEntry->m_bufOriginalImage.CopyBuffer(pBuf);
-
-        const auto index = m_vecImages.AddToTail(pEntry);
-        m_dictImages.Insert(pPath, index);
-    }
-
-    void RemoveImageFromCache(const char *pImagePath)
-    {
-        const auto index = m_dictImages.Find(pImagePath);
-        if (!m_dictImages.IsValidIndex(index))
-            return;
-
-        m_vecImages.PurgeAndDeleteElement(m_dictImages[index]);
-        m_dictImages.RemoveAt(index);
-    }
-
-    ImageCacheEntry *FindImageByPath(const char *pPath)
-    {
-        const auto index = m_dictImages.Find(pPath);
-        if (m_dictImages.IsValidIndex(index))
-        {
-            return m_vecImages[m_dictImages[index]];
-        }
-
-        return nullptr;
-    }
-
-    CUtlVector<ImageCacheEntry *> m_vecImages;
-    CUtlDict<int, uint16> m_dictImages;
-};
-
 CFileImageCache g_FileImageCache;
+CFileImageCache *g_pFileImageCache = &g_FileImageCache;
 
 FileImage::FileImage(IImage *pDefaultImage /* = nullptr*/) : m_iX(0), m_iY(0), m_iDesiredWide(0),
                                                              m_iDesiredTall(0), m_iRotation(0),
